add capture moves(without actually capturing)

+++ src/routes/games.rs
@@ -1,10 +1,10 @@
-use std::collections::HashMap;
 use askama_axum::IntoResponse;
 use axum::{
     extract::{Path, Query},
     routing::{get, post},
     Router,
 };
+use std::collections::HashMap;
 
 use crate::{store::Store, templates::BoardTemplate, utility::Point};
 
@@ -22,11 +22,12 @@ impl GamesRouter {
 
     async fn get_legal_moves(
         Path(id): Path<String>,
-        Query(mut query): Query<HashMap<String, usize>>,
+        Query(mut query): Query<HashMap<String, i8>>,
     ) -> impl IntoResponse {
         let room = Store::get_room(id).unwrap();
         let x = query.remove("x").unwrap();
         let y = query.remove("y").unwrap();
+        // let new_board = room.board.with_captures(Point { x, y });
         let new_board = room.board.with_legal_moves(Point { x, y });
         Store::update_board(room.id.clone(), new_board.clone());
         BoardTemplate {
@@ -37,7 +38,7 @@ impl GamesRouter {
 
     async fn make_move(
         Path(id): Path<String>,
-        Query(mut query): Query<HashMap<String, usize>>,
+        Query(mut query): Query<HashMap<String, i8>>,
     ) -> impl IntoResponse {
         let room = Store::get_room(id).unwrap();
         let x = query.remove("x").unwrap();
